add more oop by creating menu classes
