+++ src/main/java/seedu/bigpp/menu/Menu.java
@@ -1,0 +1,8 @@
+package seedu.bigpp.menu;
+
+public abstract class Menu {
+
+    public static void printMenu() {
+        return;
+    }
+}
