+++ src/main/java/seedu/bigpp/ui/UI.java
@@ -3,17 +3,20 @@ package seedu.bigpp.ui;
 import java.io.PrintStream;
 import java.util.Scanner;
 import java.util.Stack;
+import seedu.bigpp.menu.BuilderMenu;
+import seedu.bigpp.menu.ComponentMenu;
+import seedu.bigpp.menu.MainMenu;
+import seedu.bigpp.menu.PCViewerMenu;
+import seedu.bigpp.menu.TutorialMenu;
 
 public abstract class UI {
-    private static UIState uiState = UIState.MAIN_MENU;
 
     private static final String DIVIDER = "===================================================";
-
-    private static final String LOGO = "add logo here";
-
     private static final Scanner in = new Scanner(System.in);
     private static final PrintStream out = System.out;
 
+    public static final String LOGO = "add logo here";
+    private static UIState uiState = UIState.MAIN_MENU;
     public static Stack<UIState> visitedMenusStack = new Stack<UIState>();
 
     public static void initializeVisitedMenusStack() {
@@ -34,23 +37,19 @@ public abstract class UI {
 
         switch (uiState) {
         case MAIN_MENU:
-            if (isStartingMainMenu()) {
-                printStartingMainMenu();
-            } else {
-                printMainMenu();
-            }
+            MainMenu.printMenu();
             break;
         case VIEWER:
-            printViewer();
+            PCViewerMenu.printMenu();
             break;
         case BUILDER:
-            printBuilder();
+            BuilderMenu.printMenu();
             break;
         case TUTORIAL:
-            printTutorial();
+            TutorialMenu.printMenu();
             break;
         case COMPONENT:
-            printComponent();
+            ComponentMenu.printMenu();
             break;
         default:
             break;
@@ -59,66 +58,8 @@ public abstract class UI {
         out.println(DIVIDER);
     }
 
-    private static boolean isStartingMainMenu() {
-        return visitedMenusStack.size() == 1;
-    }
-
     public static void clearTerminal() {
         out.println("\033[H\033[2J");
-    }
-
-    public static void printStartingMainMenu() {
-        out.println(LOGO);
-        out.println("Welcome to BigPP!");
-        out.println("What would you like to do?");
-        out.println("PC viewer");
-        out.println("PC tutorial");
-        out.println("exit");
-    }
-
-    public static void printMainMenu() {
-        out.println(LOGO);
-        out.println("Welcome to BigPP!");
-        out.println("What would you like to do?");
-        out.println("PC viewer");
-        out.println("PC tutorial");
-        out.println("back");
-        out.println("exit");
-    }
-
-    public static void printViewer() {
-        out.println("PC viewer");
-        out.println("What would you like to do?");
-        out.println("View PC");
-        out.println("Build PC");
-        out.println("back");
-        out.println("exit");
-    }
-
-    public static void printBuilder() {
-        out.println("PC builder");
-        out.println("What would you like to do?");
-        out.println("Add component");
-        out.println("Remove component");
-        out.println("View PC");
-        out.println("back");
-        out.println("exit");
-    }
-
-    public static void printTutorial() {
-        out.println("PC tutorial");
-        out.println("What would you like to do?");
-        out.println("View tutorial");
-        out.println("back");
-        out.println("exit");
-    }
-
-    public static void printComponent() {
-        out.println("Component");
-        out.println("What would you like to do?");
-        out.println("View component");
-        out.println("back");
-        out.println("exit");
     }
 
     public static void showResult(String result) {
